fix: move BaseProviderConfig to impl package and cleanup logic to avoid type casting
