+++ testing/camunda-process-test-langchain4j/src/main/java/io/camunda/process/test/impl/judge/BedrockChatModelAdapterProvider.java
@@ -15,7 +15,6 @@
  */
 package io.camunda.process.test.impl.judge;
 
-import io.camunda.process.test.api.judge.BaseProviderConfig;
 import io.camunda.process.test.api.judge.ChatModelAdapter;
 import io.camunda.process.test.api.judge.ChatModelAdapterProvider;
 import io.camunda.process.test.api.judge.ProviderConfig;
@@ -24,7 +23,7 @@ public class BedrockChatModelAdapterProvider implements ChatModelAdapterProvider
 
   @Override
   public String getProviderName() {
-    return ProviderConfig.PROVIDER_AMAZON_BEDROCK;
+    return BaseProviderConfig.PROVIDER_AMAZON_BEDROCK;
   }
 
   @Override
