+++ testing/camunda-process-test-langchain4j/src/main/java/io/camunda/process/test/impl/judge/OpenAiChatModelAdapterProvider.java
@@ -15,7 +15,6 @@
  */
 package io.camunda.process.test.impl.judge;
 
-import io.camunda.process.test.api.judge.BaseProviderConfig;
 import io.camunda.process.test.api.judge.ChatModelAdapter;
 import io.camunda.process.test.api.judge.ChatModelAdapterProvider;
 import io.camunda.process.test.api.judge.ProviderConfig;
@@ -24,7 +23,7 @@ public class OpenAiChatModelAdapterProvider implements ChatModelAdapterProvider 
 
   @Override
   public String getProviderName() {
-    return ProviderConfig.PROVIDER_OPENAI;
+    return BaseProviderConfig.PROVIDER_OPENAI;
   }
 
   @Override
