+++ testing/camunda-process-test-langchain4j/src/main/java/io/camunda/process/test/impl/judge/OpenAiCompatibleChatModelAdapterProvider.java
@@ -15,7 +15,6 @@
  */
 package io.camunda.process.test.impl.judge;
 
-import io.camunda.process.test.api.judge.BaseProviderConfig;
 import io.camunda.process.test.api.judge.ChatModelAdapter;
 import io.camunda.process.test.api.judge.ChatModelAdapterProvider;
 import io.camunda.process.test.api.judge.ProviderConfig;
@@ -24,7 +23,7 @@ public class OpenAiCompatibleChatModelAdapterProvider implements ChatModelAdapte
 
   @Override
   public String getProviderName() {
-    return ProviderConfig.PROVIDER_OPENAI_COMPATIBLE;
+    return BaseProviderConfig.PROVIDER_OPENAI_COMPATIBLE;
   }
 
   @Override
